Simplify code base
