--- src/IgnoreList.ts
@@ -1,3 +1,0 @@
-export interface IgnoreList {
-  [locale: string]: string[];
-}
