+++ src/shouldIgnore.ts
@@ -1,6 +1,6 @@
-import { IgnoreList } from '../IgnoreList';
+import { IgnoreList } from '.';
 
-export function shouldIgnore(ignore: IgnoreList, locale: string, domain: string) {
+export function shouldIgnore({ ignore, locale, domain }: { ignore: IgnoreList; locale: string; domain: string }): boolean {
   for (let [ignoreLocale, ignoreDomains] of Object.entries<string[]>(ignore)) {
     if (locale === ignoreLocale && ignoreDomains.includes(domain)) {
       return true;
